@@ -1,16 +1,15 @@
 name: flutter_libphonenumber
 description: Leverages libphonenumber to allow for asynchronous and synchronous formatting of phone numbers in Flutter apps.
-version: 0.3.12
+version: 1.0.3
 homepage: https://github.com/bottlepay/flutter_libphonenumber
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.10.0"
+  sdk: ">=2.12.0-0 <3.0.0"
+  flutter: ^1.10.0
 
 dependencies:
   flutter:
     sdk: flutter
-  devicelocale: ^0.3.1
 
 dev_dependencies:
   flutter_test:
