@@ -1,8 +1,3 @@
-import 'dart:developer';
-import 'dart:io';
-
-import 'package:devicelocale/devicelocale.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter_libphonenumber/flutter_libphonenumber.dart';
 
 /// Manages countries by code and name
@@ -11,15 +6,15 @@ class CountryManager {
   factory CountryManager() => _instance;
   CountryManager._internal();
 
-  List<CountryWithPhoneCode> _countries = [];
-  String deviceLocaleCountryCode;
+  var _countries = <CountryWithPhoneCode>[];
   var _initialized = false;
 
   /// List of all supported countries on the device with phone code metadata
   List<CountryWithPhoneCode> get countries => _countries;
 
-  Future<void> loadCountries(
-      {Map<String, CountryWithPhoneCode> overrides}) async {
+  Future<void> loadCountries({
+    Map<String, CountryWithPhoneCode> overrides = const {},
+  }) async {
     if (_initialized) {
       return;
     }
@@ -29,19 +24,10 @@ class CountryManager {
           await FlutterLibphonenumber().getAllSupportedRegions();
 
       /// Apply any overrides to masks / country data
-      (overrides ?? {}).forEach((key, value) {
+      overrides.forEach((key, value) {
         phoneCodesMap[key] = value;
         // print('Applied override for $key');
       });
-
-      /// Get the device locale
-      try {
-        final locale = await Devicelocale.currentLocale;
-        deviceLocaleCountryCode = locale.substring(locale.length - 2);
-      } catch (e) {
-        // print('Error detecting deviceLocaleCountryCode, setting default GB');
-        deviceLocaleCountryCode = 'GB';
-      }
 
       /// Save list of the countries
       _countries = phoneCodesMap.values.toList();
@@ -55,17 +41,17 @@ class CountryManager {
 
 class CountryWithPhoneCode {
   CountryWithPhoneCode({
-    @required this.phoneCode,
-    @required this.countryCode,
-    @required this.exampleNumberMobileNational,
-    @required this.exampleNumberFixedLineNational,
-    @required this.phoneMaskMobileNational,
-    @required this.phoneMaskFixedLineNational,
-    @required this.exampleNumberMobileInternational,
-    @required this.exampleNumberFixedLineInternational,
-    @required this.phoneMaskMobileInternational,
-    @required this.phoneMaskFixedLineInternational,
-    @required this.countryName,
+    required this.phoneCode,
+    required this.countryCode,
+    required this.exampleNumberMobileNational,
+    required this.exampleNumberFixedLineNational,
+    required this.phoneMaskMobileNational,
+    required this.phoneMaskFixedLineNational,
+    required this.exampleNumberMobileInternational,
+    required this.exampleNumberFixedLineInternational,
+    required this.phoneMaskMobileInternational,
+    required this.phoneMaskFixedLineInternational,
+    required this.countryName,
   });
 
   /// GB locale, useful for dummy values
@@ -112,80 +98,80 @@ class CountryWithPhoneCode {
   /// ```
   /// 07400 123456
   /// ```
-  final String exampleNumberMobileNational;
+  final String? exampleNumberMobileNational;
 
   /// Example fixed line number in national format.
   /// ```
   /// 0121 234 5678
   /// ```
-  final String exampleNumberFixedLineNational;
+  final String? exampleNumberFixedLineNational;
 
   /// Phone mask for mobile number in national format.
   /// ```
   /// 00000 000000
   /// ```
-  final String phoneMaskMobileNational;
+  final String? phoneMaskMobileNational;
 
   /// Phone mask for fixed line number in national format.
   /// ```
   /// 0000 000 0000
   /// ```
-  final String phoneMaskFixedLineNational;
+  final String? phoneMaskFixedLineNational;
 
   /// Example mobile number in international format.
   /// ```
   /// +44 7400 123456
   /// ```
-  final String exampleNumberMobileInternational;
+  final String? exampleNumberMobileInternational;
 
   /// Example fixed line number in international format.
   /// ```
   /// +44 121 234 5678
   /// ```
-  final String exampleNumberFixedLineInternational;
+  final String? exampleNumberFixedLineInternational;
 
   /// Phone mask for mobile number in international format.
   /// ```
   /// +00 0000 000000
   /// ```
-  final String phoneMaskMobileInternational;
+  final String? phoneMaskMobileInternational;
 
   /// Phone mask for fixed line number in international format.
   /// ```
   /// +00 000 000 0000
   /// ```
-  final String phoneMaskFixedLineInternational;
+  final String? phoneMaskFixedLineInternational;
 
   /// Country name
   /// ```
   /// United Kingdom
   /// ```
-  final String countryName;
+  final String? countryName;
 
   @override
   String toString() =>
       '[CountryWithPhoneCode(countryName: $countryName, regionCode: $countryCode, phoneCode: $phoneCode, exampleNumberMobileNational: $exampleNumberMobileNational, exampleNumberFixedLineNational: $exampleNumberFixedLineNational, phoneMaskMobileNational: $phoneMaskMobileNational, phoneMaskFixedLineNational: $phoneMaskFixedLineNational, exampleNumberMobileInternational: $exampleNumberMobileInternational, exampleNumberFixedLineInternational: $exampleNumberFixedLineInternational, phoneMaskMobileInternational: $phoneMaskMobileInternational, phoneMaskFixedLineInternational: $phoneMaskFixedLineInternational)]';
 
-  static CountryWithPhoneCode getCountryDataByPhone(String phone,
-      {int subscringLength}) {
+  static CountryWithPhoneCode? getCountryDataByPhone(String phone,
+      {int? subscringLength}) {
     if (phone.isEmpty) return null;
     subscringLength = subscringLength ?? phone.length;
 
     if (subscringLength < 1) return null;
     var phoneCode = phone.substring(0, subscringLength);
 
-    var rawData = CountryManager().countries.firstWhere(
-        (data) => toNumericString(data.phoneCode.toString()) == phoneCode,
-        orElse: () => null);
-    if (rawData != null) {
-      return rawData;
+    try {
+      return CountryManager()
+          .countries
+          .firstWhere((data) => toNumericString(data.phoneCode) == phoneCode);
+    } on StateError catch (_) {
+      return getCountryDataByPhone(phone, subscringLength: subscringLength - 1);
     }
-    return getCountryDataByPhone(phone, subscringLength: subscringLength - 1);
   }
 
   /// Get the phone mask based on number type and format
   getPhoneMask(
-      {@required PhoneNumberFormat format, @required PhoneNumberType type}) {
+      {required PhoneNumberFormat format, required PhoneNumberType type}) {
     if (format == PhoneNumberFormat.international &&
         type == PhoneNumberType.mobile) {
       return phoneMaskMobileInternational;
