@@ -1,23 +1,1 @@
-import 'package:flutter/services.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_libphonenumber/flutter_libphonenumber.dart';
-
-void main() {
-  // const MethodChannel channel = MethodChannel('flutter_libphonenumber');
-
-  // TestWidgetsFlutterBinding.ensureInitialized();
-
-  // setUp(() {
-  //   channel.setMockMethodCallHandler((MethodCall methodCall) async {
-  //     return '42';
-  //   });
-  // });
-
-  // tearDown(() {
-  //   channel.setMockMethodCallHandler(null);
-  // });
-
-  // test('getPlatformVersion', () async {
-  //   expect(await FlutterLibphonenumber.platformVersion, '42');
-  // });
-}
+void main() {}
